+++ crates/actors/src/supervision/adapter/mod.rs
@@ -6,4 +6,5 @@ mod supervisor;
 pub use proxy::{Proxy, ProxySends, ProxySendsWithParent, ProxyWithParent};
 pub use supervisor::{
     ChildTopology, RestartConfiguration, Supervisor, SupervisorError, SupervisorSends,
+    SupervisorWithParent,
 };